refactor: Change graphql types path
